+++ src/pretalx/frontend/schedule-editor/src/lib/i18n.js
@@ -11,12 +11,12 @@ import i18next from 'i18next'
 import moment from 'moment-timezone'
 
 const localeModules = import.meta.glob('../../../../locale/*/LC_MESSAGES/django.po')
-const momentLocaleModules = import.meta.glob('../../node_modules/moment/dist/locale/*.js')
+const momentLocaleModules = import.meta.glob('../../../node_modules/moment/dist/locale/*.js')
 
 export default async function (locale) {
 	const localeModule = await localeModules[`../../../../locale/${locale}/LC_MESSAGES/django.po`]?.()
 	const momentLocale = locale.split("_")[0]
-	const momentLocaleModule = await momentLocaleModules[`../../node_modules/moment/dist/locale/${momentLocale}.js`]?.()
+	const momentLocaleModule = await momentLocaleModules[`../../../node_modules/moment/dist/locale/${momentLocale}.js`]?.()
 	moment.locale(momentLocale)
 
 	return {